Move stratum to a separate module
