+++ ethereum/blockcreation/src/test/java/org/hyperledger/besu/ethereum/blockcreation/EthHashMiningCoordinatorTest.java
@@ -51,7 +51,7 @@ public class EthHashMiningCoordinatorTest {
     assertThat(miningCoordinator.isMining()).isFalse();
     assertThat(miningCoordinator.hashesPerSecond()).isEqualTo(Optional.empty());
     assertThat(miningCoordinator.getWorkDefinition()).isEqualTo(Optional.empty());
-    assertThat(miningCoordinator.submitWork(solution.getNonce())).isFalse();
+    assertThat(miningCoordinator.submitWork(solution)).isFalse();
   }
 
   @Test
