+++ ethereum/stratum/build.gradle
@@ -1,0 +1,44 @@
+/*
+ * Copyright ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+apply plugin: 'java-library'
+
+jar {
+  baseName 'besu-ethereum-stratum'
+  manifest {
+    attributes(
+      'Specification-Title': baseName,
+      'Specification-Version': project.version,
+      'Implementation-Title': baseName,
+      'Implementation-Version': calculateVersion()
+      )
+  }
+}
+
+dependencies {
+  api project(':plugin-api')
+  api project(':util')
+
+  implementation project(':ethereum:core')
+
+  implementation 'com.google.guava:guava'
+  implementation 'io.vertx:vertx-core'
+
+  testImplementation project(':testutil')
+
+  testImplementation 'org.assertj:assertj-core'
+  testImplementation 'com.fasterxml.jackson.core:jackson-databind'
+  testImplementation 'junit:junit'
+}
