+++ settings.gradle
@@ -35,6 +35,7 @@ include 'ethereum:referencetests'
 include 'ethereum:retesteth'
 include 'ethereum:rlp'
 include 'ethereum:trie'
+include 'ethereum:stratum'
 include 'metrics:core'
 include 'metrics:rocksdb'
 include 'nat'
